+++ ui-lib/src/main/java/co/electriccoin/zcash/di/ViewModelModule.kt
@@ -78,6 +78,7 @@ import co.electriccoin.zcash.ui.screen.swap.detail.SwapDetailVM
 import co.electriccoin.zcash.ui.screen.swap.detail.support.SwapSupportVM
 import co.electriccoin.zcash.ui.screen.swap.info.SwapRefundAddressInfoVM
 import co.electriccoin.zcash.ui.screen.swap.lock.EphemeralLockVM
+import co.electriccoin.zcash.ui.screen.swap.mismatch.SwapQuoteMismatchVM
 import co.electriccoin.zcash.ui.screen.swap.orconfirmation.ORSwapConfirmationVM
 import co.electriccoin.zcash.ui.screen.swap.picker.SwapAssetPickerVM
 import co.electriccoin.zcash.ui.screen.swap.picker.SwapBlockchainPickerVM
@@ -178,6 +179,7 @@ val viewModelModule =
         viewModelOf(::ORSwapConfirmationVM)
         viewModelOf(::SwapDetailVM)
         viewModelOf(::SwapSupportVM)
+        viewModelOf(::SwapQuoteMismatchVM)
         viewModelOf(::SwapRefundAddressInfoVM)
         viewModelOf(::ScreenTimeoutVM)
         viewModelOf(::EphemeralHotfixVM)